rename pages dir to pager and rename pager components in a regular fashion
