+++ frontend/src/components/reader/pager/pdf-doc.vue
