+++ frontend/src/components/reader/pager/scale-for-scroll.vue
